+++ src/value/numeric.rs
@@ -1,8 +1,8 @@
-use super::{Number, Rational, Unit};
+use super::{Number, Rational, Unit, UnitSet};
 use crate::output::{Format, Formatted};
 use crate::Error;
 use std::fmt::{self, Display};
-use std::ops::Neg;
+use std::ops::{Div, Mul, Neg};
 
 /// A Numeric value is a [`Number`] with a [`Unit`] (which may be
 /// Unit::None).
@@ -11,7 +11,7 @@ pub struct Numeric {
     /// The number value of this numeric.
     pub value: Number,
     /// The unit of this numeric.
-    pub unit: Unit,
+    pub unit: UnitSet,
 }
 
 impl Numeric {
@@ -19,17 +19,17 @@ impl Numeric {
     ///
     /// The value can be given as anything that can be converted into
     /// a [`Number`], e.g. an [`isize`], a [`Rational`], or a [`f64`].
-    pub fn new(value: impl Into<Number>, unit: Unit) -> Numeric {
+    pub fn new<V: Into<Number>, U: Into<UnitSet>>(value: V, unit: U) -> Self {
         Numeric {
             value: value.into(),
-            unit,
+            unit: unit.into(),
         }
     }
     /// Create a new numeric value with no unit.
     pub fn scalar(value: impl Into<Number>) -> Numeric {
         Numeric {
             value: value.into(),
-            unit: Unit::None,
+            unit: UnitSet::scalar(),
         }
     }
 
@@ -43,7 +43,21 @@ impl Numeric {
     /// assert_eq!(inch.as_unit(Unit::Deg), None);
     /// ```
     pub fn as_unit(&self, unit: Unit) -> Option<Number> {
-        self.unit.scale_to(&unit).map(|scale| &self.value * &scale)
+        self.unit
+            .scale_to_unit(&unit)
+            .map(|scale| &self.value * &scale)
+    }
+    /// Convert this numeric value to a given unit, if possible.
+    ///
+    /// # Examples
+    /// ```
+    /// # use rsass::value::{Numeric, Unit};
+    /// let inch = Numeric::new(1, Unit::In);
+    /// assert_eq!(inch.as_unit(Unit::Mm).unwrap() * 5, 127.into());
+    /// assert_eq!(inch.as_unit(Unit::Deg), None);
+    /// ```
+    pub fn as_unitset(&self, unit: &UnitSet) -> Option<Number> {
+        self.unit.scale_to(unit).map(|scale| &self.value * &scale)
     }
 
     /// Convert this numeric value to a given unit, if possible.
@@ -69,7 +83,7 @@ impl Numeric {
 
     /// Return true if this value has no unit.
     pub fn is_no_unit(&self) -> bool {
-        self.unit == Unit::None
+        self.unit.is_none()
     }
 
     /// Get a reference to this `Value` bound to an output format.
@@ -91,7 +105,7 @@ impl PartialOrd for Numeric {
     fn partial_cmp(&self, other: &Numeric) -> Option<std::cmp::Ordering> {
         if self.unit == other.unit {
             self.value.partial_cmp(&other.value)
-        } else if let Some(scaled) = other.as_unit(self.unit.clone()) {
+        } else if let Some(scaled) = other.as_unitset(&self.unit) {
             self.value.partial_cmp(&scaled)
         } else {
             None
@@ -109,9 +123,30 @@ impl Neg for &Numeric {
     }
 }
 
+impl Div for &Numeric {
+    type Output = Numeric;
+    fn div(self, rhs: Self) -> Self::Output {
+        Numeric {
+            value: &self.value / &rhs.value,
+            unit: &self.unit / &rhs.unit,
+        }
+    }
+}
+impl Mul for &Numeric {
+    type Output = Numeric;
+    fn mul(self, rhs: Self) -> Self::Output {
+        Numeric {
+            value: &self.value * &rhs.value,
+            unit: &self.unit * &rhs.unit,
+        }
+    }
+}
+
 impl<'a> Display for Formatted<'a, Numeric> {
     fn fmt(&self, out: &mut fmt::Formatter) -> fmt::Result {
-        self.value.value.format(self.format).fmt(out)?;
-        self.value.unit.fmt(out)
+        let mut unit = self.value.unit.clone();
+        let value = &self.value.value * &unit.simplify();
+        value.format(self.format).fmt(out)?;
+        unit.fmt(out)
     }
 }