+++ tests/spec/libsass/units/conversion/mod.rs
@@ -4,7 +4,6 @@ use super::rsass;
 
 // From "sass-spec/spec/libsass/units/conversion/angle.hrx"
 #[test]
-#[ignore] // wrong result
 fn angle() {
     assert_eq!(
         rsass(
@@ -117,7 +116,6 @@ fn angle() {
 
 // From "sass-spec/spec/libsass/units/conversion/frequency.hrx"
 #[test]
-#[ignore] // wrong result
 fn frequency() {
     assert_eq!(
         rsass(
@@ -158,7 +156,6 @@ fn frequency() {
 
 // From "sass-spec/spec/libsass/units/conversion/resolution.hrx"
 #[test]
-#[ignore] // wrong result
 fn resolution() {
     assert_eq!(
         rsass(
@@ -229,7 +226,6 @@ fn resolution() {
 
 // From "sass-spec/spec/libsass/units/conversion/size.hrx"
 #[test]
-#[ignore] // wrong result
 fn size() {
     assert_eq!(
         rsass(
@@ -462,7 +458,6 @@ fn size() {
 
 // From "sass-spec/spec/libsass/units/conversion/time.hrx"
 #[test]
-#[ignore] // wrong result
 fn time() {
     assert_eq!(
         rsass(